backend/pyproject.toml
[project]
name = "pi-4termo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-ninja[swagger]>=1.5.0",
]
